// src/lib/enums.ts
export enum ModularFieldType {
  Checkbox = 'checkbox',
  Color = 'color',
  Date = 'date',
  DateTimeLocal = 'datetime-local',
  Email = 'email',
  File = 'file',
  Hidden = 'hidden',
  Month = 'month',
  Number = 'number',
  Password = 'password',
  Radio = 'radio',
  Search = 'search',
  Select = 'select',
  Submit = 'submit',
  Tel = 'tel',
  Text = 'text',
  Textarea = 'textarea',
  Time = 'time',
  Url = 'url',
  Week = 'week'
}

export enum FieldError {
  Empty = 'empty',
  Format = 'format',
  Group = 'group',
  Undefined = 'undefined'
}